finally seems to work
We want to put some static variables in a section .wifi_info.
For that we use a linker script memory.x. At first I tried to make the
size 256 bytes. But the initial values from the compiled program were
not being written (although there was a UF2 block for it). Inceasing the
size of the section to the flash sector size 4096 bytes seems to
work (but not if it's at the end of flash, have to test that again.) No
I'm just using an address at the far end of the address space.