fix memory layout of static data
The problem is the following:
The pico UF2 bootloader has a
bug (https://datasheets.raspberrypi.com/rp2040/rp2040-datasheet.pdf#errata-e14)
where in a certain situation, not all data might be written to the
flash.
The situation arises when a sector (4096 bytes) of the flash is partial,
i.e. the uf2 file does not include all 16 256-byte pages of the sector.
For the last sector of the binary this is okay but not for sectors in
the middle of the binary.
The Rust elf2uf2-rs utility does not include the necessary padding uf2 blocks
so I'm switching to the original elf2uf2 from the pico C sdk, which
does.
The problem only arises when I add the additional .wifi_info section because
it is disconnected from the rest of the sections and certainly results
in a partial sector. Before, the final sector containing code could be
partial, but when adding the .wifi_info section, it becomes a partial
sector in the middle of the binary.